+++ packages/browser-cli/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @msw-dev-tool/browser-cli
 
+## 2.0.2
+
+### Patch Changes
+
+- Updated dependencies [67de2dd]
+- Updated dependencies [bda8f30]
+  - @msw-dev-tool/core@1.5.1
+  - @msw-dev-tool/cli-core@0.2.3
+
+## 2.0.1
+
+### Patch Changes
+
+- Updated dependencies [a9f235c]
+  - @msw-dev-tool/core@1.5.0
+  - @msw-dev-tool/cli-core@0.2.2
+
+## 2.0.0
+
+### Patch Changes
+
+- Updated dependencies [5b505c2]
+  - @msw-dev-tool/core@1.4.0
+  - @msw-dev-tool/cli-core@0.2.1
+
 ## 1.0.0
 
 ### Minor Changes